requirements.txt
blinker==1.7.0
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.49.0
greenlet==3.0.3
gunicorn==21.2.0
html5lib==1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pycparser==2.21
pydyf==0.9.0
pyparsing==3.1.2
pyphen==0.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.28
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
webencodings==0.5.1
Werkzeug==3.0.1
zopfli==0.2.3
